// types/graphql-types.ts
export type TagQuery = { allMicrocmsPosts: { edges: Array<{ node: { id: string, description?: string | null, postsId?: string | null, publishedAt?: any | null, publishedAtOnHatena?: any | null, title?: string | null, internal: { type: string }, tags?: Array<{ color?: string | null, id?: string | null, name?: string | null } | null> | null, thumbnail?: { url?: string | null } | null } }> } };